aa_evan_airflow_tutorials/chapter10/requirements.txt
pandas
numpy
seaborn
matplotlib
requests
psycopg2-binary
streamlit
plotly
pytz
mysql-connector-python
streamlit_autorefresh
apache-airflow-providers-postgres==5.7.1